backend/requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
Flask==2.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
JSON-log-formatter==0.4.0
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-json-logger==2.0.2
requests==2.26.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.7
Werkzeug==2.0.1
